requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
#awscli
flake8
python-dotenv>=0.5.1
pep8
isort
pyarrow
pandas
hydra-core
scikit-learn
wandb
xgboost
lightgbm
numpy
optuna
imbalanced-learn
catboost
torch
torchvision
kaleido
